+++ application/public/index.php
@@ -1,0 +1,9 @@
+<?php
+
+/*
+ * This is the main front controller.
+ *
+ * https://en.wikipedia.org/wiki/Front_controller
+ */
+
+echo "Hello world!";