+++ application/public/index_dev.php
@@ -1,0 +1,11 @@
+<?php
+
+/*
+ * This check prevents access to debug front controllers that are deployed by accident to production servers.
+ */
+if (empty($_ENV['PHP_DEBUG_FRONT_CONTROLLERS_ACCESS']) || 'on' !== $_ENV['PHP_DEBUG_FRONT_CONTROLLERS_ACCESS']) {
+    header('HTTP/1.0 403 Forbidden');
+    exit('You are not allowed to access this file. Check '.basename(__FILE__).' for more information.');
+}
+
+phpinfo();