+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+# PHPdock Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -6,3 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.0.1 - 2017-08-07
+
+### Added
+
+- Nginx container with self-signed SSL certificate.
+- PHP container with Xdebug and Crontab.
+- Composer container.
+- NPM container.
+- Redis container.
+- MariaDB container.
+- PhpMyAdmin container.
+- MailHog container.
+- Storage container (data-only container pattern).